Fix annoying warnings explosion into hevcdsp filters and video_dsp files

+++ libavcodec/x86/hevcdsp.h
@@ -463,7 +463,7 @@ void oh_hevc_sao_band_filter_0_12_sse(uint8_t *_dst, uint8_t *_src, ptrdiff_t _s
            int y_BL, int x_EL, int y_EL, int block_w, int block_h, int widthEL, int heightEL,
            const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_luma_h_x2_sse_16(int16_t *dst, ptrdiff_t dststride, uint16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_luma_h_x2_sse_16(int16_t *dst, ptrdiff_t dststride, uint8_t *_src, ptrdiff_t srcstride,
                int x_EL, int x_BL, int width, int height, int widthEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
@@ -471,15 +471,15 @@ void oh_hevc_sao_band_filter_0_12_sse(uint8_t *_dst, uint8_t *_src, ptrdiff_t _s
                int x_EL, int x_BL, int block_w, int block_h, int widthEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_luma_v_x2_sse_16(uint16_t *_dst, ptrdiff_t _dststride, int16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_luma_v_x2_sse_16(uint8_t *_dst, ptrdiff_t _dststride, int16_t *_src, ptrdiff_t srcstride,
                int y_BL, int x_EL, int y_EL, int width, int height, int widthEL, int heightEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_cr_v_x2_sse_16(uint16_t *_dst, ptrdiff_t dststride, int16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_cr_v_x2_sse_16(uint8_t *_dst, ptrdiff_t dststride, int16_t *_src, ptrdiff_t srcstride,
            int y_BL, int x_EL, int y_EL, int width, int height, int widthEL, int heightEL,
            const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_luma_h_x1_5_sse_16(int16_t *dst, ptrdiff_t dststride, uint16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_luma_h_x1_5_sse_16(int16_t *dst, ptrdiff_t dststride, uint8_t *_src, ptrdiff_t srcstride,
                int x_EL, int x_BL, int width, int height, int widthEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
@@ -487,11 +487,11 @@ void oh_hevc_sao_band_filter_0_12_sse(uint8_t *_dst, uint8_t *_src, ptrdiff_t _s
                int x_EL, int x_BL, int block_w, int block_h, int widthEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_luma_v_x1_5_sse_16(uint16_t *_dst, ptrdiff_t _dststride, int16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_luma_v_x1_5_sse_16(uint8_t *_dst, ptrdiff_t _dststride, int16_t *_src, ptrdiff_t srcstride,
                int y_BL, int x_EL, int y_EL, int width, int height, int widthEL, int heightEL,
                const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
 
-   void oh_upsample_filter_block_cr_v_x1_5_sse_16(uint16_t *_dst, ptrdiff_t dststride, int16_t *_src, ptrdiff_t srcstride,
+   void oh_upsample_filter_block_cr_v_x1_5_sse_16(uint8_t *_dst, ptrdiff_t dststride, int16_t *_src, ptrdiff_t srcstride,
            int y_BL, int x_EL, int y_EL, int width, int height, int widthEL, int heightEL,
            const struct HEVCWindow *Enhscal, struct UpsamplInf *up_info);
